Implemented Buy Transactions. (#2)

+++ ggc-core/src/ggc/BuyTransaction.java
@@ -2,7 +2,7 @@ package ggc;
 
 public class BuyTransaction extends Transaction{
 
-    public BuyTransaction(int id, int paymentDate, String productId, String partnerId, int quantity, int price){
+    public BuyTransaction(int id, int paymentDate, String productId, String partnerId, int quantity, float price){
         super(id, paymentDate, productId, partnerId, quantity, price);
     }
